Things that used to be very easy for me are now difficult, namely, using the keyboard with my right hand. I am still in the middle of my recovery, and there is no end in sight. So what I am doing right now is just some light programming to get used to the voice software.
Sure, it might be a light, but this is good practice. Yesterday I did some research into sorting methods and as I said, it's all block wide and work wide inside the libraries. So you have to get the merge sort that I implemented for the FPGA after I put it in, I found some bugs in the compiler and fixed them just now. Spiral V2.5.2 should be out soon.

It is not that I don't want to work hard, but working hard could Aggravate my injury. So I am always being careful on the side of laziness. Right now, it feels like the world is telling me to read manga and watch enemy. So that is what I will do.

I'm going to do tomorrow is check out the random number generation methods. But I don't really need those for just a hand ranker. So maybe to Morrow, i'll just put some work into porting the hand anchor on the gpu. It won't be difficult at all.

From so far it has been my vacation and for the next several months it will be like that. So let me get back to it.