I redid the hand ranker. Now it should be going from top to bottom.
I think this qualifies as a light bit of work for the day. It is good keyboard practice. Even though I had the glove 80 keyboard for three weeks now, I still haven't gotten used to it due to lack of practice with irrelevant key presses. The programming I've been doing so far like. in today's session is actually quite rare. I've only started getting back into programming.

Right there. I'm thinking what I should do next, and what I should do is do the testing. Maybe it seems simple, but it only now occurs to me that I'm not going to be able to test anything on the cpu. Instead, follow the testing is going to be on the GPU instead. What I'm going to do is compare it to the original hand drinker. And for the sake of testing, I think I'm finally going to need the random number generator.

Let me stop here so I can do the chores. And then I'll think whether I want to do it today, or leave it for tomorrow.