"11:15am. https://docs.cupy.dev/en/stable/reference/random.html
Let me do some programming today. I'm going to figure out how the random number generation is done using coupi, and how the generator states can be passed into the kernels.

I just need to focus on this. I need to figure out how the kuda standard library functions can be used. I did this in 2018 literally. So there is no reason why I couldn't be able to do this today. And we heated the physiotherapy that was assigned to me. I think he's going a senile and he isn't replying to my emails in Italian manner.

2:00 PM. Done with breakfast. Let me resume any to figure out how to run The device kernels with the random functions using coupon.

Let me look at the good examples again. I need to check out. I need to compare the host API with what I have with Cupie already.

I need to see whether there is any overlap between the host api and the device api. Ardo's bit generators, something I can use on the device api, or do I need to do it all from scratch?

https://docs.nvidia.com/cuda/curand/host-api-overview.html#host-api-overview

Academy what is the sobel algorithm?

https://youtu.be/wGflWqY41RU
Brian Hayes: Orderly Randomness - Quasirandom Numbers and Quasi-Monte Carlo | IACS Seminar

https://youtu.be/wGflWqY41RU?t=1044

These stories about random numbers are actually pretty interesting. It seems they're real random numbers are correlated and biased, and generally inferior to the fakes.

https://youtu.be/wGflWqY41RU?t=1608
> It turns out there were tiny biases in the random number generator.

Honestly, I thought this talk would be a snooze fest, but I am quite enjoying it.

https://youtu.be/wGflWqY41RU?t=2989

Here he finally mentions this a Sobol pattern.

https://youtu.be/wGflWqY41RU?t=3196

This is very informative. I didn't know at all that the Monte Carlo integration using pseudo and the quasi random numbers has a different error rate. It seems a pseudo random numbers are something one should be using in high dimensions.

https://docs.nvidia.com/cuda/curand/device-api-overview.html#device-api-overview

After some thinking, I would change my mind. It seems that I'm not going to be able to use anything. I mean, all of these device api structures on the python side. Instead, I'm going to have to initialize them directly on the device, but that is fine actually.

///

```c
__device__ unsigned int
curand(curandStatePhilox4_32_10_t *state)
```

Following a call to curand_init(), curand() returns a sequence of pseudorandom numbers with a period
. If curand() is called with the same initial state each time, and the state is not modified between the calls to curand(), the same sequence is always generated.

```c
__device__ void
curand_init(unsigned long long seed,
            unsigned long long subsequence,
            unsigned long long offset,
            curandStatePhilox4_32_10_t *state)
```

///

I just need to call the `curand_init` to init the state on the device and the I will have my pick of the random functions after that.

No problem.

That is all that I needed to know. In the API example they use global memory to store the state, but I think that doing it in local memory should be fine.

4:15pm. Made an appointment with a physiotherapeutist. I think I am going to call it a day here. I don't think my hands will be able to bear any more typing.

I think just figuring out how the curand_init functions work is enough for today.

In the past, there were plenty of days such as this. There is where I just studied and and taught and did research, Instead of doing programming. At some point, I think I'm actually going to have to figure out how to do the Python bindings to C classes, or maybe C classes, C tracks. I mean, but that day isn't today and I don't care about it. As far as I think concerned, my name injured or sick. The 6:00 PM rule doesn't apply. But to make up for it, I'm going to go back and check out some of those web 3 posts, but that an anonymous person.

///

>>97820433
If you have to ask such a simple question that you should be googling instead then you won't make it anon.

However I will play along:

> What is a Blockchain: /watch?v=kHybf1aC-jE
> What are Smart Contracts: /watch?v=pyaIppMhuic
> What is Web 3: /watch?v=nHhAEkG1y2U

Keep in mind pretty much every example in these videos above are basic analogies and the space has already evolved in the 2 years that most of the videos explaining what crypto is were made. You would be better off reading blogs on medium for individual deep dives.

The truth is though, crypto has lots of filtering layers. If you want to be a dev in it then exclusively only read and learn about smart contracts and the EVM and Ethereum. Ignore every single thing else. No bitcoin, no NFTs, no Solana, no memecoin, etc...

> Solidity 101: /watch?v=kdvVwGrV7ec
> Best Solidity playlist on yt: /playlist?list=PLO5VPQH6OWdVQwpQfw9rZ67O6Pjfo6q-p

Best book on Smart Contracts, Ethereum and the EVM: https://github.com/ethereumbook/ethereumbook

It's free on github but I have a physical copy and recommend you get one too as a beginner.

Other than the above, you really need to live in this space. There is a reason why roasties and normies hate on crypto because anyone who is involved in it from the gamblers to the investors to the developers all have to eat sleep crypto and to an outsider that seems really culty. But the truth is that is the only way to deal with the massive information flow that occurs on a daily basis. Everything moves so fast that you don't have time to get bored.

///

I was on the one he was talking to in that thread, but he caught my attention and basically know nothing about blockchain except that it's a scam. So maybe I should educate myself to see if there is anything useful in that field. It's not like I don't have the time."