"2:50pm. Let me do some more programming. What I want to do right now is implement the division method as it is even more elegant.
I didn't understand it at first, but I do now.

https://www.pcg-random.org/posts/bounded-rands.html

Ah, whatever. Let me implement the integer multiplication method since I am at it. I am having a bit too much fun studying this code.

```cpp
uint32_t bounded_rand(rng_t& rng, uint32_t range) {
    // calculates 2**32 % range
    uint32_t t = (-range) % range;
    for (;;) {
        uint32_t r = rng();
        if (r >= t)
            return r % range;
    }
}
```

Let me go with this as it is the easiest to understand.

4:30 PM.

```spiral
inl uint_range forall t {number; uint} . rand {from nearTo} s =
    inl range : t = nearTo - from
    // calculates 2**32 % range
    inl bottom = (0 - range) % range
    let rec loop() =
        inl r = rand s
        if r >= bottom then r else loop()
    loop() % range + from
```

I had a bit too much fun with the integer samplers today. My hand is really paying the price for it. It's feeling weak and I cannot possibly do anymore programming. I'm going to stop it here. This was a pretty interesting exercise. I learned quite a bit about samling integers. I didn't know about the division and the modular samplers. The method by Lemire, which uses just multiplication, is interesting, but I didn't want to bring it in because it's not that much faster and it's a lot more complicated. For the method that I have here, it's very easy to understand, actually, once you walk through it on a simplified examples. I think this module of one is the simplest of the branch. And once the compiler inlines the constants, it will be very fast, much faster than the bitmask one that I implemented originally.

I could blame myself for taking too much time on this simple issue, but when you're programming on your own, taking the time to really think about what you are doing is one of the big benefits that makes actually programming fun. It is hard to get better if you are just rushing through things without thinking them through."