I think the work that I just did just now should take care of the reconnects after the computer goes to sleep. I overdid it, and I think I strained my right hand a bit too much. I'll take a break here.
This is one of the issues with the rsi. It's actually hard to stop. when you start to feel the strain. You really want to at least take care of the bug that you are working on.

I think I am being foolish. In the future, when I start to feel the strain, I should just stop period and do nothing else. For half an hour at least.