"12:45pm. Did some work today and my hand held up remarkable despite me working for nearly two hours. I'm not sure whether I should do a bit more work today on my own project. And Italy should be resting as much as possible.
2:40 PM. Don't read the breakfast. I want to do some work right now, but I really like my hand not hurting. So at most I'm going to do some voice programming.

Before I do that, let me post the absence of regional article on the similar chrome blog.

Cut crap, Wordpress really sucks right now.

3:00 PM. It turns out that the ad blocker was causing the issue. Well, to be more precise, the site was blocking the air blocker.

Forget manga. Let me finally do some programming here. Last time I did a random player. And now I'm going to begin the end. What I'm going to do here is going to be my ultimate project. It is what is going to be the basis for the singularity.

I'm going to back to 2018 and combine it with my work from 2021. I'm going to bring all of my experiences together into one singular program.

3:50 PM. An obstacle before me is the Cutlass library. It is a really integral part of building a GPU only. ML library. I'm going to have to take the time in order to go through the docks for it.

https://docs.nvidia.com/cuda/cuda-c-programming-guide/index.html#spatial-partitioning-also-known-as-warp-specialization

I really don't understand why you would want to leave one warp hanging at a barrier.

https://github.com/NVIDIA/cutlass/blob/main/media/docs/fundamental_types.md

This basic Array is a much better than the one that I am using.

4:40 PM. This library is a real beast. It has a ton of documentation, but it isn't very useful to me right now.

I spent a while immersing myself in it. But the reality is that I am going to have to play with it.

For the next few sessions, I'm going to dedicate my time to investigating the Qatlas Library.

And that's going said now that I come this far, I'm modeling whether there are any pure GPU ML libraries. For all I know, there might be some.

After a brief search, it doesn't seem like it. If there were, I would definitely be able to find them. At this point in time, implementing everything on the gpu doesn't seem to be a popular pastime.

Nevermind that. Tomorrow, I'm going to download the git repo for Cutlass, and then try using some of the functions inside. The goal will be to figure out how to use the metrics multiplication kernels inside another one."