"9:40am. I got a message that the trackpad will get here today. Let me read the Akuma Koujo and after I catch up, I'll resume studying the Cutlass examples.
10:50am. The delivery came! Right now I am trying it out. My first experience is quite positive. It is very smooth and precise. It's a lot more precise than the **** Ella come left handed mouse.

I'm just wondering where the controls for it are.

11:10 AM. I guess I had to wait a bit for the touchpad settings to appear in Windows search. It started working much better after I installed the driver.

11:45 AM. https://content-security-policy.com/strict-dynamic/

> An allowlist for specific inline scripts using a cryptographic nonce (number used once). The server must generate a unique nonce value each time it transmits a policy. It is critical to provide an unguessable nonce, as bypassing a resource's policy is otherwise trivial. See unsafe inline script for an example. Specifying nonce makes a modern browser ignore 'unsafe-inline' which could still be set for older browsers without nonce support.

I am starting to get a bit interested in security. I didn't know that specifying a nonce disables the unsafe inline.

https://github.com/NVIDIA/cutlass/blob/8236f30675bbe98f81d11c05764b77bfcb25b8cc/examples/12_gemm_bias_relu/gemm_bias_relu.cu#L93

It seems that this library can fuse matrix multiplication and bias addition with a RELU operation.

1:00 PM. I'm really enjoying the magic trackpad. It's a lot better than the Kensington expert at any rate. I really wasted way too much money on these **** track balls. As expected, I should have just gone with the trackpad from the start. You know what? Now that I got this, I should just cancel that 20 euro left handed vertical mouse.

I'll do that first thing after breakfast.

...Actually, I did it just now. That takes care of that.

Let me take a break here.

4:05 PM. And this was a long break and I wanted to be even longer. Let me cheer for a bit more, and then I'll go back to studying the examples. I still haven't found any of the kernel api ones.

To be honest, just reading the examples is draining my motivation to go on. I honestly don't care about matrix multiplication all that much. The library feels very autistic. But at the same time, that is what is making me trust it. A lot of work went into it.

4:40 PM. https://github.com/NVIDIA/cutlass/blob/main/examples/37_gemm_layernorm_gemm_fusion/gemm_with_layernorm.h

This **** is so **** complicated. Even a single example would take me days to understand that completely.

It's all a very ugly C++ code.

https://github.com/NVIDIA/cutlass/blob/main/examples/41_fused_multi_head_attention/kernel_forward.h

Seriously, what the hell am I supposed to do with this?

https://github.com/NVIDIA/cutlass/issues/

Opened issue 1291 here to ask a question.

5:40 PM. It's annoying, but I'm going to need some help from the authors in order to figure out how to proceed. The examples are incredibly complex.

https://github.com/NVIDIA/cutlass/blob/8236f30675bbe98f81d11c05764b77bfcb25b8cc/include/cutlass/gemm/kernel/gemm.h#L203

Maybe this is what I should be using?

6:05 PM. `cutlass::gemm::kernel`

And starting to get, I think this is, in fact, what I should be using. I see now that it's getting cold on the device. But it has a **** load of parameters that I don't know what they are supposed to be for.

Anyway, this single insight is helping me focus. I'll just do a search through the repo for the uses of this struct.

You know what? Forget about a matrix multiply. Let me just start things off with trying to get a simple dot product working.

After that, I'll move from there. These C++ libraries are no joke.

Now that I've done a bit of studying, the best thing to do is to just get a started with coding.

Staring at the examples isn't going to get me anywhere, but making mistakes and fixing them as I go along well.

6:15pm. I'll close it here for the day."