"11:25am. https://developer.nvidia.com/blog/new-cublas-12-0-features-and-matrix-multiplication-performance-on-nvidia-hopper-gpus/
> The cuBLAS library is an implementation of Basic Linear Algebra Subprograms (BLAS) on top of the NVIDIA CUDA runtime, and is designed to leverage NVIDIA GPUs for various matrix multiplication operations. This post mainly discusses the new capabilities of the cuBLAS and cuBLASLt APIs. However, the cuBLAS library also offers cuBLASXt API targeting single-node multiGPU GEMMs. The cuBLASDx API is set to be available in Early Access in 2023 and targets GEMMs and their fusion inside device functions.

11:35 AM. Okay, forget about this. I got up really late today, but never minded that. It feels like I am getting lazier and lazier.

In my defense, it's harder not to be so when the hand is in the state, it is currently.

Let me do my morning reading. I'll have breakfast. Take a bath and then I'll do some programming.

The usual.

https://publicdomainreview.org/collection/books-fatal-to-their-authors/

2:50 PM. https://docs.cupy.dev/en/stable/reference/generated/cupy.ndarray.html

I've been thinking what I should do Today. Yesterday I got the cooperative group lunches understood and what I should do to day is get started to work with on multidimensional tensors. Right now, I only have arrays, and I'm going to have to figure out how to allocate tensors themselves.

Once I have that, I'll be able to start to work on the machine learning library.

I haven't gotten approved for that early access matter library, but it's Saturday and tomorrow it will be Sunday, so I doubt it will happen that early. I'll probably get approved the next week. I doubt I will get rejected for this.

Let me just do some experimentation in python so I can figure out what the strides for multidimensional Cooper arrays are.

```py
import cupy as cp
x = cp.random.random((2,4,15),cp.float32)
print(x.strides) # (240, 60, 4)
```

It doesn't seem like it's doing anything special to pad them.

https://stackoverflow.com/questions/58905574/how-to-allocate-pitched-2d-memory-in-cupy

3:50 PM. This is such a pain in the **** And actually, you have to go into look at my old code from 2018. To think I'd ever study that again. 2021 was fake.

It is 2024 that is going to be the real continuation to the story of 2018.

The inverse arrays that I have right now simply aren't enough. I need the multi dimensional Tensors.

4:40 PM. Done with chores. Let me take a look at my old code."