"10:20 AM. I won't paste the entire thing here, but it is bloody brilliant right now.
You know there's something I want to do. I'm going to make it so that the default ints and floats are parsed based on what is in the environment. I've been thinking about this for a while.

No, that wouldn't be a good feature. Never mind that for now. I guess this is good enough. Let me commit it. The next thing I should do is actually track out.

Actually, you know what? Let me in fact change the default in Saint floats depending on what is in the core based library. I changed my mind again. But they come in mind here is the easiest solution."