"11:15am. I'm going to do some programming here today. Yesterday I got a fever, but now I'm a bit better. I got some proper sleep. Let's see if the latest modifications that I did to the language worked. I really meant it there to be a way to change the default ins and floats.
Anything wanted to do it in any way, but the one that I did, but all the other ways had blockers. I could have passed the default int and floated types in the forlalls, but that wouldn't have worked for type literals. In the end, this was the only way left to me.

Let's see if it works. Also, this has been a long time coming, but what I want to do right now is turn off those **** information logs.

```fs
// Since Spiral v2.6.0 the default int and floats are set in the plugin settings.
// The two types are also provided based on that by the language itself.

// type int = i32
// type float = f32

// Having 32 bit ints and floats is a good setting for Cuda related backends as Nvidia GPUs do not have hardware
// support for 64 bit int arithmetic, and limited support for 64 bit floats.

// In truth, I didn't want these settings to be passed in during startup, but all the other ways had severe flaws.
// Passing the `int` and `float` through the type system in foralls wouldn't have worked for type literals for example.
// Having the settings be backend specific wouldn`t have meshed with providing editor support.
// Having the defaults be specified in the package.spiproj files would result in modules packaegs having separate defaults.

// Specifying them in the plugin settings was the best choice in the end.
```

The int and float defaults feature that I worked on last time was nearly flawless. It will be out with the upcoming version of Spiral. The next are tensor tests.

12:05 PM. Let me take a short break first. This week I will be on a sick leave. It's either one thing or another with me. First, it was my hands, and now it's Covid again."