"11:25 AM. I had breakfast so let me do some programming today. The joints on my right hand hurt, but it's not the usual numbness. So I think I'll be able to do some programming today.
I won't to do too much because I am still ill from Covid.

The main goal is to try to get the matrix multiply to work.

If I could do that, that would be a remarkable accomplishment. Now let's dive into it. I'll see whether I can do anything done today.

https://www.ycombinator.com/companies/wasmer/jobs

I kinda want to try applying here, but I'm not in the condition to. Sigh.

Let's focus on the task at hand. I need to focus on that instead of my greed.

1:15pm.

```
open corebase
open corecuda
open tensorm

inl main() =
    inl grid_range () : int = $"gridDim.x * blockDim.x"
    inl linear_id () : int = $"threadIdx.x + blockIdx.x * blockDim.x"

    inl random_normal dim =
        inl len : int = real open real_core in tensorm.utils.foldBack (*) dim 1
        inl t : array f32 = $"cp.random.normal(0,1,!len,cp.float32)"
        fromArray t |> reshape dim

    inl cp_matmul (a : tensor (int * int) float) (b : tensor (int * int) float) : tensor (int * int) float =
        inl a_dim, b_dim = a.dim, b.dim
        inl a_body,b_body : array float * array float = real tensorm.utils.map (fun (tensor_body {array}) => array) (a.bodies, b.bodies)
        $"cp.matmul(!a_body.reshape(!a_dim),!b_body.reshape(!b_dim)).flatten()" |> fromArray |> reshape (fst a_dim, snd b_dim)

    inl [a; b] = listm.map random_normal ([5, 10; 10, 5] : _ (int * int))
    inl c = cp_matmul a b
    a, b, c
    // inl blocks = 1
    // inl grids = 1 // divup (length out) blocks
    // run grids blocks (fun () =>
    //     ()
    //     )
```

Here is what I did so far. I am going to stop here for the day. I cannot go on anymore. The fever is making me light headed and I feel exhausted.

It is not even about the hands right now. And simply need some rest.

I've pushed v2.6.1 out, so let me stop here.

Now that I have said everything up for the next session, once it comes, I'll be able to start work on trying out the device matrix multiply from Cutlass. Getting that to work will be a pain in the ass, but once I clear that hurdle, nothing will stop me from starting work on the new ML library. It will be a rehash of 2018 except this time, I will have all the tools and the hardware that I need.

I am looking forward to it."