10:15am. It's another day, and I had a lot of time to think about what I want to do today. I'm going to stop doing that exercise that I've been doing for the past week because I think I've been overdoing it and the joint on my right hand needs time to heal.
That's what the screencast I really should stop it, but I can't. I'm going to translate the examples, and then I'm going to rebuild them using catalase. It seems, after all, that living it aside leaves us sour taste in my mouth.

That's why I'm going to go all the way.

For now, let me chill, and then I'll get started.

12:30 PM. It is not going well. I'm getting lost in the code. Let me show a breakfast here, and then I'll resume.

12:55 PM. Let me save the link here.

https://github.com/NVIDIA/cuda-samples/blob/e8568c417356f7e66bb9b7130d6be7e55324a519/Samples/3_CUDA_Features/tf32TensorCoreGemm/tf32TensorCoreGemm.cu#L579C17-L579C37

I prepared the file for the translation of this and brought in all the constants.

Honestly, if I was doing this on my own, it wouldn't be too hard, but doing it live is a bit tricky. I need to explore, and I'm not exactly sure what I need to do.

Let me take a break here.

https://inference-review.com/article/the-man-who-carried-computer-science-on-his-shoulders
> It seems Dijkstra never applied for any grants—though he did receive at least one, to employ a PhD student—nor did he bring any money, in the form of research contracts, to the institutions he worked for. He also never purchased a computer. Eventually, in the late 1980s, he was given one as a gift by a computer company, but never used it for word processing. Dijkstra did not own a TV, a VCR, or even a mobile phone. He preferred to avoid the cinema, citing an oversensitivity to visual input. By contrast, he enjoyed attending classical music concerts.

Based.

2:45 PM. Darwin breakfast. Let me chill for a bit, and then I'll do the chores. After that, I'm not sure whether I'd want to do any more programming, because my right hand and its rest. And think I'll spend some time studying the examples.

4:15 PM. I really want to do some more programming, even though I shouldn't.

Let me do it.

4:50 PM. My pace is that of a snail, but I'm going to stop here in order to let my hand relax.

5:00 PM. It is really fun thinking about this. But let me go back to chilling. I guess doing a little bit every day is enough.

I envy the past days where I could program for as long as I wanted.