https://www.cs.utexas.edu/users/EWD/transcriptions/EWD12xx/EWD1284.html > After more than 45 years in the field, I am still convinced that in computing, elegance is not a dispensable luxury but a quality that decides between success and failure;
https://news.ycombinator.com/item?id=39120687
> In this respect the programmer does not differ from any other craftsman: unless he loves his tools it is highly improbable that he will ever create something of superior quality.

I love this last quote. I only heard of it just now.

https://www.youtube.com/watch?v=mEcYjBwSA8Y
3-HOUR Progressive Metal and Djent Mix – Aggressive / Instrumental / Melodic

I like this playlist.

1/27/2024

12:15pm. I've gone completely insane and then end up thinking all night about matrix multiplication and convolution. As a result, I didn't get a wink of sleep, but I got many strong insights about how to implement these two things. Also, that exercise that I've been praising for the past week really agitated my joint and my hand hurt during the night. So I am going to take a break for the next few days. Maybe I made a mistake by doing that exercise. It could have made things worse.

But I did expect it to make things worse before making it better. So who knows? We'll see. It is worth seeing if the rebound after the first rest period is an improvement.

Tonight was absolutely crazy. I just had to know how to do those things and my brain refused to go offline.

I think I should be fine from here on out.

1/28/2024

12:35 PM. Today is a day for resting. My right hand is still in a bad shape to not as bad as yesterday. I think I might want to do some voice programming today.

6:25 PM. I did some good programming today using my voice. Let me find that I have a lunch.

It is satisfying to do a little bit every day. I'll continue like this tomorrow.