8:45am. https://manhuascan.com/manga/26606-kenja-no-deshi-wo-nanoru-kenja-official/chapter-3
It turns out the official scans have a pretty good quality. The line work is much better.

9:55 AM. I made a physiotherapy appointment tomorrow at 2:30 PM. I've been waiting to schedule this for 4 weeks. It was a stroke of luck that I managed to do this today.

Right now I should be able to resume programming. Let me just finish the chapter that I'm reading, and then I'll go for it.

https://developer.nvidia.com/nsight-compute-videos

I guess I'll be watching this today. I might as well take a break from programming.

https://youtu.be/z5bbTi__GTk?t=1888
e Seminar #22 Tools and techniques for making efficient use of GPUs

According to this, it's possible to go from the GPU directly to the File. system without going over the CPU.

This is something new to me. I didn't know this existed.

https://youtu.be/z5bbTi__GTk?t=2658

Uh, so that's how it works. I didn't know that inside graphics existed.

https://youtu.be/z5bbTi__GTk?t=2772

In retrospect, it was pretty stupid of Maine not to use these tools back in 2021.

But I can imagine very well that I would have seen that the performance is ****

I didn't need the profiling tools to tell me that. And she has didn't have a way of fixing that back then, not without dishing the pie torch library and taking the approach that I am taking now.

https://youtu.be/z5bbTi__GTk?t=4560

It seems Nsight Compute has a Python API.

https://youtu.be/z5bbTi__GTk?t=5880

This is quite amazing. I didn't expect the cuda tooling to be that good. This will be quite useful.

https://youtu.be/z5bbTi__GTk?t=6156

Actually, I didn't know the SASS was available.

I thought that was private information to the GPU vendor.

https://tenstorrent.com/research/unboxing-the-tenstorrent-grayskull-ai-accelerator/

It seems tense torrent is finally releasing something. Let me check out this Youtube video.

The product page says it is apparently priced at $600. But it's inferred only, so it's useless to me.

https://youtu.be/z5bbTi__GTk?t=7164
> Until Linux side, we have debuggers called Cuda Gdb

He also mentioned the compute sanitizer.

> This is actually the result of porn. The normal reaction to seeing a woman you desire fuck another man is supposed to be disgust. Porn has conditioned people to now be attracted to that exact dynamic. What follows is the desire for the very women men were supposed to naturally avoid, which are promiscuous and slutty woman who will align more towards increased sexual deviance.

Actually never thought of it that way. You read something good. Everyone's in a vial on fortune.

https://docs.tenstorrent.com/tenstorrent/v/tt-buda/advanced_user_guide#comparison-to-gpu-programming-model

These guys finally have the ducks.

///

Pitfalls/Oddities

Register Spilling
The compiler does not implement register spilling. Since Grayskull only has 4 LRegs, running out of registers is a common occurence. If you see the following: error: cannot store SFPU register (reigster spill?) - exiting! you have most likely run out of registers.

Error Messages
Unfortunately, many errors are attributed to the code in the wrapper rather than in the code being written. For example, using an unitialized variable would show an error at a macro called by a wrapper function before showing the line number in the user’s code.

Function Calls
There is no abi and none of the vector types can be passed on the stack. Therefore, all function calls must be inlined. To ensure this use sfpi_inline, which is defined to __attribute__((always_inline)) on GCC.

Unnecessary Moves
The gcc compiler occasionally moves a value from one register to another for no apparent reason. At this point it appears there is nothing that can be done about this besides hoping that the issue is fixed in a future version of gcc.

///

Yeah, it turns out that really happened insane to expect anything. They should have been at this level 5 years ago.

I don't think I'll be implementing the poker game on this device. Is going to be Gpus all the way. TT in 2030 might come up with something useful.

Much like a **** **** trading during my first 6 years of doing it, it seems I've also **** **** my programming.

I'll live and learn from this. Maybe in the next life I'll avoid it.

https://www.nvidia.com/en-us/on-demand/session/gtcspring22-s41723/

At the 7:16 he is going into the relationship between shared memory and the L1 cache.

At 12:15 it seems like this video is actually a pretty good Tutorial on shared memory.

5:25 PM. I very confused why L1 cache doesn't have the same access pattern as shared memory.

https://www.nvidia.com/en-us/on-demand/session/gtcspring22-s41723/

Let me watch the optics stuff for a bit.

7:05 PM. https://youtu.be/Z7QsPb7YWjc?t=1604

This is pretty boring, so let me stop here for the day.

I'll think about here and then catch up with Kenja No Deshi.

It feels like I wasted my entire day, especially if you compare it to those intense programming sessions over the past week. But I'm going to dig into it tomorrow and figure out how to profile and analyze cuda kernels.