11:05am. I am done for the day.
My even my left hand right now is too strained to do anything. I simply have to stop a programming.

It seems that I finally gotten to the point where both of my hands are too messed up to do anything.

I ran into a weird compiler bug with the CUDA back end.

I guess I'll watch an enemy and the read novels for the rest of the day.