This is horrible. I went through the full benchmarking ground for all the examples and none of them beat the stoch matrix multiply. Despite it not using tensor cores.
In fact, they're barely better than the naive matrix multiply in the first example.

This is horrible.

To make matters even worse, the final example actually stops working once I turn off the asserts. It's complete ****

Tomorrow, I'm going to have to do a summary of what is going on. And then I'm going to have to debug what the problem in the final example is. Once I figured that out, I'll crack it open in inside compute and see what the bottlenecks in the kernel are.

I guess it was too much to hope for that everything would workout right away. I'll have to keep on struggling until I get it right.