And here it is. This is our action selector.
It can be used to sample from a discrete probability distribution.

This is a lot more efficient that you would get with the pie torch because all the kernel steps are used into one here and all the computations are done in registers.

We are going to be using this for all the activation functions, as well as layer normalization, softmax and sampling.

It might not seem like a big deal now that we've done all this.

But if I could show this to My past self, he would have been deeply impressed.

We made this look easy.