"10:20am. https://www.reddit.com/r/reinforcementlearning/comments/1by4xtx/how_difficult_is_it_to_train_dqns_for_toy_marl/
///

I try to train them for a few hundred thousand steps, and usually arrive at a point where they (seem to?) converge to a Nash equilibrium, with games ending in a tie. Except that when I try running either of them against a random opponent, they still lose some 10% of the time, which means they haven't learned the optimum strategy.

///

///

Yes, it is very hard to solve even seemingly simple MARL environments with RL techniques, and DQN is not meant for that at all, especially if you are using a replay buffer.

The problem does not come from hyperparameter tuning or lack of exploration, it comes from the high, adversarial non-stationarity of multi-agent learning environments, which breaks the fundamental assumptions of single-agent RL. Your agents are not learning to play against random opponents, they are learning to play against themselves (if you are using self-play), which is an adversarily moving target in terms of the underlying learning process.

Trying to tackle MARL problems is a common mistake for beginners, as the theory behind it is extremely involved and there is no way one can understand what goes wrong in a multi-agent setting without a strong understanding of the single-agent setting.

///

I'm going to showcase this video in the series.

https://github.com/luchris429/purejaxrl

///

PureJaxRL is a high-performance, end-to-end Jax Reinforcement Learning (RL) implementation. When running many agents in parallel on GPUs, our implementation is over 1000x faster than standard PyTorch RL implementations. Unlike other Jax RL implementations, we implement the entire training pipeline in JAX, including the environment. This allows us to get significant speedups through JIT compilation and by avoiding CPU-GPU data transfer. It also results in easier debugging because the system is fully synchronous. More importantly, this code allows you to use jax to jit, vmap, pmap, and scan entire RL training pipelines. With this, we can:

///

>🦎 Discover new RL algorithms with meta-evolution

https://chrislu.page/blog/meta-disco/

https://arxiv.org/abs/2111.05803
Gradients are Not All You Need

Done with breakfast. I was going to immediately start Programming today, but I say think into the future I don't actually have a algorithm that I want to implement for evolutionary learning.

The research direction in that blog post really looks interesting to me. I want to read some of these papers.

Let's leave it for a bit. I'm not really motivated right now to program. I'll do that a bit later in the day."