```powershell # ffmpeg -i input.tta -c:a libmp3lame -q:a 2 output.mp3 # ffmpeg -i input.tta -c:a flac output.flac
$files = Get-ChildItem '.\`[彩音 ～xi-on～`]\**\*.tta'
foreach ($file in $files){
    $output_path = [System.IO.Path]::ChangeExtension($file.FullName,"flac")
    if ([System.IO.Path]::Exists($output_path) -eq $false) {
        ffmpeg -i $file.FullName -c:a flac $output_path
    }
}
```

Here is a little script to convert TTA files to FLAC ones.

I am starting to appreciate Powershell more and more lately.

Right now the three videos are rendering and should be posted to Youtube tomorrow. They are are 10 1/2 hours in total.

Not as much as the last batch, but it's still a sizeable.

I finally done with work for the day and it's time for some well deserved the rest. Tomorrow, if I'm lucky, I'll be able to start work on the next video.